+++ src/components/PhotoCard/index.js
@@ -1,50 +1,20 @@
-import React, { useEffect, useState, useRef } from 'react'
+import React from 'react'
 
 import { Article, ImgWrapper, Img, Button } from './styles'
 import { MdFavoriteBorder, MdFavorite } from 'react-icons/md'
+import { useLocalStorage } from '../../Hooks/useLocalStorage'
+import { useNearScreen } from "../../Hooks/useNearScreen"
+
 const DEFAULT_IMAGE = 'https://images.unsplash.com/photo-1518791841217-8f162f1e1131?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=800&q=60'
 
 export const PhotoCard = ({
   id, likes = 0, src = DEFAULT_IMAGE
 }) => {
-  const element = useRef(null)
-  const [Show, setShow] = useState(false)
-
-  useEffect(function () {
-    Promise.resolve(
-      typeof window.IntersectionObserver !== 'undefined'
-        ? window.IntersectionObserver : import('intersection-observer')
-    ).then(() => {
-      const observer = new window.IntersectionObserver(
-        function (entries) {
-          const { isIntersecting } = entries[0]
-          if (isIntersecting) {
-            setShow(true)
-            observer.disconnect()
-          }
-        })
-      observer.observe(element.current)
-    })
-  }, [element])
+  const [Show, element] = useNearScreen()
 
   const key = `like-${id}`
-  const [Liked, setLiked] = useState(() => {
-    try {
-      const like = window.localStorage.getItem(key)
-      return JSON.parse(like)
-    } catch (e) {
-      return console.log(e)
-    }
-  })
+  const [Liked, setLiked] = useLocalStorage(key, false)
   const Icon = Liked ? MdFavorite : MdFavoriteBorder
-  const setLocalStorage = value => {
-    try {
-      window.localStorage.setItem(key, value)
-      setLiked(value)
-    } catch (e) {
-      console.error(e)
-    }
-  }
 
   return (
     <Article ref={element}>
@@ -56,7 +26,7 @@ export const PhotoCard = ({
                 <Img src={src} />
               </ImgWrapper>
             </a>
-            <Button onClick={() => setLocalStorage(!Liked)}>
+            <Button onClick={() => setLiked(!Liked)}>
               <Icon size='32px' />
               {likes} likes!
             </Button>
